+++ tests/tests_pytorch/deprecated_api/test_remove_2-0.py
@@ -18,6 +18,7 @@ import pytest
 
 import pytorch_lightning
 from pytorch_lightning import Callback, Trainer
+from pytorch_lightning.cli import LightningCLI
 from pytorch_lightning.demos.boring_classes import BoringModel
 from tests_pytorch.callbacks.test_callbacks import OldStatefulCallback
 from tests_pytorch.helpers.runif import RunIf
@@ -320,3 +321,12 @@ def test_v2_0_0_unsupported_on_init_start_end(callback_class, tmpdir):
         RuntimeError, match="callback hook was deprecated in v1.6 and is no longer supported as of v1.8"
     ):
         trainer.validate(model)
+
+
+@pytest.mark.parametrize(
+    ["name", "value"],
+    [("description", "description"), ("env_prefix", "PL"), ("env_parse", False)],
+)
+def test_lightningCLI_parser_init_params_deprecation_warning(name, value):
+    with mock.patch("sys.argv", ["any.py"]), pytest.deprecated_call(match=f".*{name!r} init parameter is deprecated.*"):
+        LightningCLI(BoringModel, run=False, **{name: value})