+++ tests/tests_pytorch/test_cli.py
@@ -231,7 +231,7 @@ def test_lightning_cli_args(tmpdir):
 def test_lightning_env_parse(tmpdir):
     out = StringIO()
     with mock.patch("sys.argv", ["", "fit", "--help"]), redirect_stdout(out), pytest.raises(SystemExit):
-        LightningCLI(BoringModel, BoringDataModule, env_parse=True)
+        LightningCLI(BoringModel, BoringDataModule, parser_kwargs={"default_env": True})
     out = out.getvalue()
     assert "PL_FIT__CONFIG" in out
     assert "PL_FIT__SEED_EVERYTHING" in out
@@ -246,7 +246,7 @@ def test_lightning_env_parse(tmpdir):
         "PL_FIT__TRAINER__LOGGER": "false",
     }
     with mock.patch.dict(os.environ, env_vars), mock.patch("sys.argv", ["", "fit"]):
-        cli = LightningCLI(BoringModel, BoringDataModule, env_parse=True)
+        cli = LightningCLI(BoringModel, BoringDataModule, parser_kwargs={"default_env": True})
     assert cli.config.fit.data.data_dir == str(tmpdir)
     assert cli.config.fit.trainer.default_root_dir == str(tmpdir)
     assert cli.config.fit.trainer.max_epochs == 1
@@ -535,6 +535,16 @@ def test_cli_config_overwrite(tmpdir):
         LightningCLI(BoringModel, trainer_defaults=trainer_defaults)
     with mock.patch("sys.argv", argv):
         LightningCLI(BoringModel, save_config_kwargs={"overwrite": True}, trainer_defaults=trainer_defaults)
+
+
+def test_cli_config_filename(tmpdir):
+    with mock.patch("sys.argv", ["any.py", "fit"]):
+        LightningCLI(
+            BoringModel,
+            trainer_defaults={"default_root_dir": str(tmpdir), "logger": False, "max_steps": 1, "max_epochs": 1},
+            save_config_kwargs={"config_filename": "name.yaml"},
+        )
+    assert os.path.isfile(tmpdir / "name.yaml")
 
 
 @pytest.mark.parametrize("run", (False, True))